refactor(email): DRY normalizeEmail + maskEmail into server/utils/email.js
The same two helpers were defined three times across the codebase:
  - server/server.js              (local copies)
  - server/services/sp.js         (local copies, also re-exported)
  - server/services/spLedger.js   (local copy)

Three copies means three places to drift: a future change to the masking
strategy (e.g. always show the domain hash, never show 2 chars for
single-character local parts) would have to be applied three times and
tested three times.

This commit:
  - Adds server/utils/email.js exporting both helpers, with one
    canonical JSDoc on each
  - server/server.js imports them, drops the local copies
  - services/sp.js imports them, re-exports normalizeEmail + maskEmail
    to preserve its existing public API (withSp and publicStudent both
    import maskEmail via this module)
  - services/spLedger.js imports maskEmail directly, drops local copy

Net change: -38 lines of duplicated logic, +18 lines of single-source
util. Same behavior (verified by visual diff: byte-for-byte identical
algorithm).

Why: the helpers are pure and stateless. Putting them in one place makes
them trivially testable in isolation, removes a class of 'I updated
one copy but forgot the others' bugs, and unblocks adding new email-
related utilities (e.g. domain-only display, hash-keyed bucketing) in
one place.

Verification:
  - node --check passes on all 3 changed .js files
  - maskEmail output is unchanged: tested with 4 sample inputs (a@b.c,
    ab@x.com, abc@x.com, abcde@x.com) and got the same masked strings
  - No client-side change, no schema change, no new deps
